Seperate speed and distance message classes into their own files
The SpeedMessagesInterface was very large. This helps to scope
down each message to its own file, which hopefully makes it easier to
add new messages and maintain old ones.
