+++ isobus/CMakeLists.txt
@@ -30,11 +30,14 @@ set(ISOBUS_SRC
     "nmea2000_fast_packet_protocol.cpp"
     "isobus_data_dictionary.cpp"
     "isobus_language_command_interface.cpp"
+    "isobus_machine_selected_speed.cpp"
+    "isobus_machine_selected_speed_command.cpp"
     "isobus_time_date_interface.cpp"
     "isobus_task_controller_client_objects.cpp"
     "isobus_task_controller_client.cpp"
     "isobus_device_descriptor_object_pool.cpp"
     "isobus_shortcut_button_interface.cpp"
+    "isobus_ground_based_speed.cpp"
     "isobus_functionalities.cpp"
     "isobus_guidance_interface.cpp"
     "isobus_speed_distance_messages.cpp"
@@ -45,6 +48,7 @@ set(ISOBUS_SRC
     "isobus_heartbeat.cpp"
     "isobus_task_controller_server.cpp"
     "isobus_task_controller_server_options.cpp"
+    "isobus_wheel_based_speed.cpp"
     "nmea2000_message_definitions.cpp"
     "nmea2000_message_interface.cpp"
     "isobus_device_descriptor_object_pool_helpers.cpp"
