+++ test/speed_distance_message_tests.cpp
@@ -127,18 +127,18 @@ TEST(SPEED_MESSAGE_TESTS, SpeedMessages)
 		EXPECT_TRUE(testPlugin.get_queue_empty());
 
 		EXPECT_EQ(63, interfaceUnderTest.machineSelectedSpeedTransmitData.get_exit_reason_code());
-		EXPECT_EQ(SpeedMessagesInterface::MachineSelectedSpeedData::LimitStatus::NotAvailable, interfaceUnderTest.machineSelectedSpeedTransmitData.get_limit_status());
-		EXPECT_EQ(SpeedMessagesInterface::MachineDirection::NotAvailable, interfaceUnderTest.machineSelectedSpeedTransmitData.get_machine_direction_of_travel());
+		EXPECT_EQ(MachineSelectedSpeedData::LimitStatus::NotAvailable, interfaceUnderTest.machineSelectedSpeedTransmitData.get_limit_status());
+		EXPECT_EQ(MachineDirection::NotAvailable, interfaceUnderTest.machineSelectedSpeedTransmitData.get_machine_direction_of_travel());
 		EXPECT_EQ(0, interfaceUnderTest.machineSelectedSpeedTransmitData.get_machine_distance());
 		EXPECT_EQ(0, interfaceUnderTest.machineSelectedSpeedTransmitData.get_machine_speed());
 		EXPECT_EQ(nullptr, interfaceUnderTest.machineSelectedSpeedTransmitData.get_sender_control_function());
 		EXPECT_EQ(0, interfaceUnderTest.machineSelectedSpeedTransmitData.get_timestamp_ms());
-		EXPECT_EQ(SpeedMessagesInterface::MachineSelectedSpeedData::SpeedSource::NotAvailable, interfaceUnderTest.machineSelectedSpeedTransmitData.get_speed_source());
+		EXPECT_EQ(MachineSelectedSpeedData::SpeedSource::NotAvailable, interfaceUnderTest.machineSelectedSpeedTransmitData.get_speed_source());
 
-		EXPECT_EQ(SpeedMessagesInterface::WheelBasedMachineSpeedData::ImplementStartStopOperations::NotAvailable, interfaceUnderTest.wheelBasedSpeedTransmitData.get_implement_start_stop_operations_state());
-		EXPECT_EQ(SpeedMessagesInterface::WheelBasedMachineSpeedData::KeySwitchState::NotAvailable, interfaceUnderTest.wheelBasedSpeedTransmitData.get_key_switch_state());
-		EXPECT_EQ(SpeedMessagesInterface::WheelBasedMachineSpeedData::OperatorDirectionReversed::NotAvailable, interfaceUnderTest.wheelBasedSpeedTransmitData.get_operator_direction_reversed_state());
-		EXPECT_EQ(SpeedMessagesInterface::MachineDirection::NotAvailable, interfaceUnderTest.wheelBasedSpeedTransmitData.get_machine_direction_of_travel());
+		EXPECT_EQ(WheelBasedMachineSpeedData::ImplementStartStopOperations::NotAvailable, interfaceUnderTest.wheelBasedSpeedTransmitData.get_implement_start_stop_operations_state());
+		EXPECT_EQ(WheelBasedMachineSpeedData::KeySwitchState::NotAvailable, interfaceUnderTest.wheelBasedSpeedTransmitData.get_key_switch_state());
+		EXPECT_EQ(WheelBasedMachineSpeedData::OperatorDirectionReversed::NotAvailable, interfaceUnderTest.wheelBasedSpeedTransmitData.get_operator_direction_reversed_state());
+		EXPECT_EQ(MachineDirection::NotAvailable, interfaceUnderTest.wheelBasedSpeedTransmitData.get_machine_direction_of_travel());
 		EXPECT_EQ(0, interfaceUnderTest.wheelBasedSpeedTransmitData.get_machine_distance());
 		EXPECT_EQ(0, interfaceUnderTest.wheelBasedSpeedTransmitData.get_timestamp_ms());
 		EXPECT_EQ(0, interfaceUnderTest.wheelBasedSpeedTransmitData.get_machine_speed());
@@ -149,25 +149,25 @@ TEST(SPEED_MESSAGE_TESTS, SpeedMessages)
 		EXPECT_EQ(0, interfaceUnderTest.groundBasedSpeedTransmitData.get_timestamp_ms());
 		EXPECT_EQ(0, interfaceUnderTest.groundBasedSpeedTransmitData.get_machine_speed());
 		EXPECT_EQ(nullptr, interfaceUnderTest.groundBasedSpeedTransmitData.get_sender_control_function());
-		EXPECT_EQ(SpeedMessagesInterface::MachineDirection::NotAvailable, interfaceUnderTest.groundBasedSpeedTransmitData.get_machine_direction_of_travel());
+		EXPECT_EQ(MachineDirection::NotAvailable, interfaceUnderTest.groundBasedSpeedTransmitData.get_machine_direction_of_travel());
 	}
 
 	{
 		TestSpeedInterface interfaceUnderTest(testECU, false, false, true, false); // Configure MSS to be sent only
 
 		interfaceUnderTest.machineSelectedSpeedTransmitData.set_exit_reason_code(15);
-		interfaceUnderTest.machineSelectedSpeedTransmitData.set_limit_status(SpeedMessagesInterface::MachineSelectedSpeedData::LimitStatus::LimitedLow);
-		interfaceUnderTest.machineSelectedSpeedTransmitData.set_machine_direction_of_travel(SpeedMessagesInterface::MachineDirection::Forward);
+		interfaceUnderTest.machineSelectedSpeedTransmitData.set_limit_status(MachineSelectedSpeedData::LimitStatus::LimitedLow);
+		interfaceUnderTest.machineSelectedSpeedTransmitData.set_machine_direction_of_travel(MachineDirection::Forward);
 		interfaceUnderTest.machineSelectedSpeedTransmitData.set_machine_distance(123456);
 		interfaceUnderTest.machineSelectedSpeedTransmitData.set_machine_speed(1000);
-		interfaceUnderTest.machineSelectedSpeedTransmitData.set_speed_source(SpeedMessagesInterface::MachineSelectedSpeedData::SpeedSource::NavigationBasedSpeed);
+		interfaceUnderTest.machineSelectedSpeedTransmitData.set_speed_source(MachineSelectedSpeedData::SpeedSource::NavigationBasedSpeed);
 
 		EXPECT_EQ(15, interfaceUnderTest.machineSelectedSpeedTransmitData.get_exit_reason_code());
-		EXPECT_EQ(SpeedMessagesInterface::MachineSelectedSpeedData::LimitStatus::LimitedLow, interfaceUnderTest.machineSelectedSpeedTransmitData.get_limit_status());
-		EXPECT_EQ(SpeedMessagesInterface::MachineDirection::Forward, interfaceUnderTest.machineSelectedSpeedTransmitData.get_machine_direction_of_travel());
+		EXPECT_EQ(MachineSelectedSpeedData::LimitStatus::LimitedLow, interfaceUnderTest.machineSelectedSpeedTransmitData.get_limit_status());
+		EXPECT_EQ(MachineDirection::Forward, interfaceUnderTest.machineSelectedSpeedTransmitData.get_machine_direction_of_travel());
 		EXPECT_EQ(123456, interfaceUnderTest.machineSelectedSpeedTransmitData.get_machine_distance());
 		EXPECT_EQ(1000, interfaceUnderTest.machineSelectedSpeedTransmitData.get_machine_speed());
-		EXPECT_EQ(SpeedMessagesInterface::MachineSelectedSpeedData::SpeedSource::NavigationBasedSpeed, interfaceUnderTest.machineSelectedSpeedTransmitData.get_speed_source());
+		EXPECT_EQ(MachineSelectedSpeedData::SpeedSource::NavigationBasedSpeed, interfaceUnderTest.machineSelectedSpeedTransmitData.get_speed_source());
 
 		ASSERT_FALSE(interfaceUnderTest.test_wrapper_send_ground_based_speed());
 		ASSERT_FALSE(interfaceUnderTest.test_wrapper_send_wheel_based_speed());
@@ -203,21 +203,21 @@ TEST(SPEED_MESSAGE_TESTS, SpeedMessages)
 	{
 		TestSpeedInterface interfaceUnderTest(testECU, false, true, false, false); // Configure wheel speed to be sent only
 
-		interfaceUnderTest.wheelBasedSpeedTransmitData.set_implement_start_stop_operations_state(SpeedMessagesInterface::WheelBasedMachineSpeedData::ImplementStartStopOperations::StartEnableImplementOperations);
-		interfaceUnderTest.wheelBasedSpeedTransmitData.set_key_switch_state(SpeedMessagesInterface::WheelBasedMachineSpeedData::KeySwitchState::NotOff);
-		interfaceUnderTest.wheelBasedSpeedTransmitData.set_machine_direction_of_travel(SpeedMessagesInterface::MachineDirection::Reverse);
+		interfaceUnderTest.wheelBasedSpeedTransmitData.set_implement_start_stop_operations_state(WheelBasedMachineSpeedData::ImplementStartStopOperations::StartEnableImplementOperations);
+		interfaceUnderTest.wheelBasedSpeedTransmitData.set_key_switch_state(WheelBasedMachineSpeedData::KeySwitchState::NotOff);
+		interfaceUnderTest.wheelBasedSpeedTransmitData.set_machine_direction_of_travel(MachineDirection::Reverse);
 		interfaceUnderTest.wheelBasedSpeedTransmitData.set_machine_distance(5000);
 		interfaceUnderTest.wheelBasedSpeedTransmitData.set_machine_speed(9876);
 		interfaceUnderTest.wheelBasedSpeedTransmitData.set_maximum_time_of_tractor_power(3);
-		interfaceUnderTest.wheelBasedSpeedTransmitData.set_operator_direction_reversed_state(SpeedMessagesInterface::WheelBasedMachineSpeedData::OperatorDirectionReversed::NotReversed);
+		interfaceUnderTest.wheelBasedSpeedTransmitData.set_operator_direction_reversed_state(WheelBasedMachineSpeedData::OperatorDirectionReversed::NotReversed);
 
-		EXPECT_EQ(SpeedMessagesInterface::WheelBasedMachineSpeedData::ImplementStartStopOperations::StartEnableImplementOperations, interfaceUnderTest.wheelBasedSpeedTransmitData.get_implement_start_stop_operations_state());
-		EXPECT_EQ(SpeedMessagesInterface::WheelBasedMachineSpeedData::KeySwitchState::NotOff, interfaceUnderTest.wheelBasedSpeedTransmitData.get_key_switch_state());
-		EXPECT_EQ(SpeedMessagesInterface::MachineDirection::Reverse, interfaceUnderTest.wheelBasedSpeedTransmitData.get_machine_direction_of_travel());
+		EXPECT_EQ(WheelBasedMachineSpeedData::ImplementStartStopOperations::StartEnableImplementOperations, interfaceUnderTest.wheelBasedSpeedTransmitData.get_implement_start_stop_operations_state());
+		EXPECT_EQ(WheelBasedMachineSpeedData::KeySwitchState::NotOff, interfaceUnderTest.wheelBasedSpeedTransmitData.get_key_switch_state());
+		EXPECT_EQ(MachineDirection::Reverse, interfaceUnderTest.wheelBasedSpeedTransmitData.get_machine_direction_of_travel());
 		EXPECT_EQ(5000, interfaceUnderTest.wheelBasedSpeedTransmitData.get_machine_distance());
 		EXPECT_EQ(9876, interfaceUnderTest.wheelBasedSpeedTransmitData.get_machine_speed());
 		EXPECT_EQ(3, interfaceUnderTest.wheelBasedSpeedTransmitData.get_maximum_time_of_tractor_power());
-		EXPECT_EQ(SpeedMessagesInterface::WheelBasedMachineSpeedData::OperatorDirectionReversed::NotReversed, interfaceUnderTest.wheelBasedSpeedTransmitData.get_operator_direction_reversed_state());
+		EXPECT_EQ(WheelBasedMachineSpeedData::OperatorDirectionReversed::NotReversed, interfaceUnderTest.wheelBasedSpeedTransmitData.get_operator_direction_reversed_state());
 
 		ASSERT_FALSE(interfaceUnderTest.test_wrapper_send_ground_based_speed());
 		ASSERT_TRUE(interfaceUnderTest.test_wrapper_send_wheel_based_speed());
@@ -255,11 +255,11 @@ TEST(SPEED_MESSAGE_TESTS, SpeedMessages)
 	{
 		TestSpeedInterface interfaceUnderTest(testECU, true, false, false, false); // Configure ground speed to be sent only
 
-		interfaceUnderTest.groundBasedSpeedTransmitData.set_machine_direction_of_travel(SpeedMessagesInterface::MachineDirection::Forward);
+		interfaceUnderTest.groundBasedSpeedTransmitData.set_machine_direction_of_travel(MachineDirection::Forward);
 		interfaceUnderTest.groundBasedSpeedTransmitData.set_machine_distance(80000);
 		interfaceUnderTest.groundBasedSpeedTransmitData.set_machine_speed(9999);
 
-		EXPECT_EQ(SpeedMessagesInterface::MachineDirection::Forward, interfaceUnderTest.groundBasedSpeedTransmitData.get_machine_direction_of_travel());
+		EXPECT_EQ(MachineDirection::Forward, interfaceUnderTest.groundBasedSpeedTransmitData.get_machine_direction_of_travel());
 		EXPECT_EQ(80000, interfaceUnderTest.groundBasedSpeedTransmitData.get_machine_distance());
 		EXPECT_EQ(9999, interfaceUnderTest.groundBasedSpeedTransmitData.get_machine_speed());
 
@@ -282,7 +282,7 @@ TEST(SPEED_MESSAGE_TESTS, SpeedMessages)
 		                                    (static_cast<std::uint16_t>(testFrame.data[5]) << 24));
 		EXPECT_EQ(80000, decodedDistance_mm);
 		EXPECT_EQ(0xFF, testFrame.data[6]);
-		EXPECT_EQ(static_cast<std::uint8_t>(SpeedMessagesInterface::MachineDirection::Forward), testFrame.data[7] & 0x03);
+		EXPECT_EQ(static_cast<std::uint8_t>(MachineDirection::Forward), testFrame.data[7] & 0x03);
 
 		// Test above max values
 		interfaceUnderTest.groundBasedSpeedTransmitData.set_machine_distance(4211081216);
@@ -296,11 +296,11 @@ TEST(SPEED_MESSAGE_TESTS, SpeedMessages)
 
 		interfaceUnderTest.machineSelectedSpeedCommandTransmitData.set_machine_selected_speed_setpoint_limit(12345);
 		interfaceUnderTest.machineSelectedSpeedCommandTransmitData.set_machine_speed_setpoint_command(56789);
-		interfaceUnderTest.machineSelectedSpeedCommandTransmitData.set_machine_direction_of_travel(SpeedMessagesInterface::MachineDirection::Forward);
+		interfaceUnderTest.machineSelectedSpeedCommandTransmitData.set_machine_direction_of_travel(MachineDirection::Forward);
 
 		EXPECT_EQ(12345, interfaceUnderTest.machineSelectedSpeedCommandTransmitData.get_machine_selected_speed_setpoint_limit());
 		EXPECT_EQ(56789, interfaceUnderTest.machineSelectedSpeedCommandTransmitData.get_machine_speed_setpoint_command());
-		EXPECT_EQ(SpeedMessagesInterface::MachineDirection::Forward, interfaceUnderTest.machineSelectedSpeedCommandTransmitData.get_machine_direction_command());
+		EXPECT_EQ(MachineDirection::Forward, interfaceUnderTest.machineSelectedSpeedCommandTransmitData.get_machine_direction_command());
 
 		ASSERT_FALSE(interfaceUnderTest.test_wrapper_send_ground_based_speed());
 		ASSERT_FALSE(interfaceUnderTest.test_wrapper_send_wheel_based_speed());
@@ -419,9 +419,9 @@ TEST(SPEED_MESSAGE_TESTS, ListenOnlyModeAndDecoding)
 		EXPECT_EQ(30, mss->get_exit_reason_code());
 		EXPECT_EQ(965742, mss->get_machine_distance());
 		EXPECT_EQ(4000, mss->get_machine_speed());
-		EXPECT_EQ(SpeedMessagesInterface::MachineSelectedSpeedData::LimitStatus::OperatorLimitedControlled, mss->get_limit_status());
-		EXPECT_EQ(SpeedMessagesInterface::MachineDirection::Forward, mss->get_machine_direction_of_travel());
-		EXPECT_EQ(SpeedMessagesInterface::MachineSelectedSpeedData::SpeedSource::GroundBasedSpeed, mss->get_speed_source());
+		EXPECT_EQ(MachineSelectedSpeedData::LimitStatus::OperatorLimitedControlled, mss->get_limit_status());
+		EXPECT_EQ(MachineDirection::Forward, mss->get_machine_direction_of_travel());
+		EXPECT_EQ(MachineSelectedSpeedData::SpeedSource::GroundBasedSpeed, mss->get_speed_source());
 		EXPECT_NE(0, mss->get_timestamp_ms());
 	}
 
@@ -457,10 +457,10 @@ TEST(SPEED_MESSAGE_TESTS, ListenOnlyModeAndDecoding)
 		auto wheelSpeed = interfaceUnderTest.get_received_wheel_based_speed(0);
 		ASSERT_NE(nullptr, wheelSpeed);
 
-		EXPECT_EQ(SpeedMessagesInterface::WheelBasedMachineSpeedData::ImplementStartStopOperations::StartEnableImplementOperations, wheelSpeed->get_implement_start_stop_operations_state());
-		EXPECT_EQ(SpeedMessagesInterface::WheelBasedMachineSpeedData::KeySwitchState::NotOff, wheelSpeed->get_key_switch_state());
-		EXPECT_EQ(SpeedMessagesInterface::WheelBasedMachineSpeedData::OperatorDirectionReversed::Reversed, wheelSpeed->get_operator_direction_reversed_state());
-		EXPECT_EQ(SpeedMessagesInterface::MachineDirection::Forward, wheelSpeed->get_machine_direction_of_travel());
+		EXPECT_EQ(WheelBasedMachineSpeedData::ImplementStartStopOperations::StartEnableImplementOperations, wheelSpeed->get_implement_start_stop_operations_state());
+		EXPECT_EQ(WheelBasedMachineSpeedData::KeySwitchState::NotOff, wheelSpeed->get_key_switch_state());
+		EXPECT_EQ(WheelBasedMachineSpeedData::OperatorDirectionReversed::Reversed, wheelSpeed->get_operator_direction_reversed_state());
+		EXPECT_EQ(MachineDirection::Forward, wheelSpeed->get_machine_direction_of_travel());
 		EXPECT_EQ(965742, wheelSpeed->get_machine_distance());
 		EXPECT_EQ(4000, wheelSpeed->get_machine_speed());
 		EXPECT_EQ(200, wheelSpeed->get_maximum_time_of_tractor_power());
@@ -500,7 +500,7 @@ TEST(SPEED_MESSAGE_TESTS, ListenOnlyModeAndDecoding)
 
 		EXPECT_EQ(965742, groundSpeed->get_machine_distance());
 		EXPECT_EQ(4000, groundSpeed->get_machine_speed());
-		EXPECT_EQ(SpeedMessagesInterface::MachineDirection::Forward, groundSpeed->get_machine_direction_of_travel());
+		EXPECT_EQ(MachineDirection::Forward, groundSpeed->get_machine_direction_of_travel());
 		EXPECT_NE(0, groundSpeed->get_timestamp_ms());
 	}
 
@@ -534,7 +534,7 @@ TEST(SPEED_MESSAGE_TESTS, ListenOnlyModeAndDecoding)
 		ASSERT_NE(nullptr, command);
 
 		EXPECT_NE(0, command->get_timestamp_ms());
-		EXPECT_EQ(SpeedMessagesInterface::MachineDirection::Reverse, command->get_machine_direction_command());
+		EXPECT_EQ(MachineDirection::Reverse, command->get_machine_direction_command());
 		EXPECT_EQ(5000, command->get_machine_selected_speed_setpoint_limit());
 		EXPECT_EQ(4000, command->get_machine_speed_setpoint_command());
 		EXPECT_NE(nullptr, command->get_sender_control_function());
